fix: Rename headers from .hpp to .h
