+++ src/libsyncengine/update_detection/file_system_observer/filesystemobserverworker.h
@@ -23,7 +23,7 @@
 #include "snapshot/livesnapshot.h"
 #include "db/syncdb.h"
 #include "libcommon/utility/types.h"
-#include "../../test/syncpal_test_helper/setinitialsituation.hpp"
+#include "../../test/syncpal_test_helper/setinitialsituation.h"
 
 #include <list>
